Create POC with hooks of do_fork, elf loading and kill
1. Remove some sample code

2. Hooking _do_fork:
It looks that several syscalls related to process creation eventually call
 __do_fork. Therefore it is better to hook it rather than individual syscalls.

3. Hooking sys_kill and __x64_sys_kill:
Newer kernels changed their prefix for syscalls there for __x64_sys_kill
should be tracked.

4. Hooking load_elf_binary
We are going to get executable name and additional information here after elf
is loaded. Some of this useful info can be VM_AREAs of the task.

Signed-off-by: Yan Vugenfirer <yan@bladerunner.io>